simulation of a pulse hitting a dielectric medium

+++ fd1d/cuda/fd1d_1_3.cu
@@ -15,7 +15,7 @@
 
 __device__
 float gaussian(int t, int t0, float sigma) {
-    return exp(-0.5*(t - t0)/sigma*(t - t0)/sigma);
+    return expf(-0.5f*(t - t0)/sigma*(t - t0)/sigma);
 }
 
 
@@ -36,7 +36,7 @@ void hyfield(int nx, float *ex, float *hy, float *bc) {
     if (idx == nx-1) ex[nx-1] = bc[3], bc[3] = bc[2], bc[2] = ex[nx-2];
     /* calculate the Hy field */
     for (int i = idx; i < nx-1; i += stx)
-        hy[i] += 0.5 * (ex[i] - ex[i+1]);
+        hy[i] += 0.5f * (ex[i] - ex[i+1]);
 }
 
 
@@ -44,7 +44,7 @@ float *dielectric(int nx, float epsr) {
     float *cb;
     cudaMallocManaged(&cb, nx*sizeof(*cb));
     for (int i = 0; i < nx; cb[i] = 0.5f, i++);
-    for (int i = nx/2; i < nx; cb[i] = 0.5/epsr, i++);
+    for (int i = nx/2; i < nx; cb[i] = 0.5f/epsr, i++);
     return cb;
 }
